feat : UI 로직을 포함하는 MainView 클래스 생성
