+++ src/main/kotlin/lotto/MainView.kt
@@ -1,0 +1,7 @@
+package lotto
+
+//UI 로직 포함하는 클래스
+class MainView {
+
+
+}
